+++ src/components/create-guild/IconSelector/index.ts
