+++ src/components/create-guild/PickRolePlatform/components/PlatformOption.tsx
