+++ src/components/create-guild/Requirements/Requirements.tsx
