+++ src/components/create-guild/Requirements/components/NftFormCard/NftFormCard.tsx
