+++ src/components/create-guild/Requirements/components/SnapshotFormCard/SnapshotFormCard.tsx
