+++ src/components/create-guild/Requirements/components/SnapshotFormCard/index.ts
