+++ src/components/create-guild/Requirements/components/Symbol.tsx
